Implement users in the database and login.
This commit adds a new table users and alters the email_subscriptions table
to a new user_subscriptions that has a foreign key to user_id instead of
email.

Also makes changes to modules/auth.js to use the users table in the
database instead of a mock hash.

And finally makes changes to server.js so that GET and POST requests
to /login.html and /logout authenticate the user and begin a session.